+++ src/main/resources/templates/ShowVillageById.html
@@ -210,7 +210,7 @@ footer p {
     <tbody>
     <th:block th:each="distance : ${T(com.ludogoriesoft.villagelifefrontend.enums.Distance).values()}">
         <tr >
-            <td colspan="2"></td> <!-- Празен <td> елемент -->
+            <td colspan="2"></td>
         </tr>
         <tr class="header-row">
             <td colspan="2">
